refactor types file to entities
